Refactor php.
If you don't care if the request is $_GET or $_POST just use $_REQUEST

Instead of a bunch of if statements around 1 variable you can use a switch statement or use each expected/allowed value as the key of an array.

Generic php objects are converted to generic javascript object via json_encode so you don't have to manually write any of the JSON syntax.